+++ clients/ios/Other Sources/InAppSettingsKit/Models/IASKSettingsStoreUserDefaults.m
@@ -74,7 +74,7 @@
     return [self.defaults doubleForKey:key];
 }
 
-- (int)integerForKey:(NSString*)key {
+- (NSInteger)integerForKey:(NSString*)key {
     return [self.defaults integerForKey:key];
 }
